use react and material-ui for the filter widget
